requirements.txt
openai
torch
PyPDF2
ollama
pyyaml
beautifulsoup4
lxml
python-dotenv
boto3
